% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_WCESurvZ.R
\name{theme_WCESurvZ}
\alias{theme_WCESurvZ}
\title{theme_WCESurvZ}
\usage{
theme_WCESurvZ(
  base_size = 12,
  base_family = "",
  font.main = c(16, "plain", "black"),
  font.submain = c(15, "plain", "black"),
  font.x = c(14, "plain", "black"),
  font.y = c(14, "plain", "black"),
  font.caption = c(15, "plain", "black"),
  font.tickslab = c(12, "plain", "black"),
  font.legend = c(10, "plain", "black"),
  ...
)
}
\arguments{
\item{base_size}{Base font size for the theme. Default is 12.}

\item{base_family}{Base font family for the theme. Default is "".}

\item{font.main}{Font settings for the main title. Default is c(16, "plain", "black").}

\item{font.submain}{Font settings for the subtitle. Default is c(15, "plain", "black").}

\item{font.x}{Font settings for the x-axis title. Default is c(14, "plain", "black").}

\item{font.y}{Font settings for the y-axis title. Default is c(14, "plain", "black").}

\item{font.caption}{Font settings for the plot caption. Default is c(15, "plain", "black").}

\item{font.tickslab}{Font settings for the axis tick labels. Default is c(12, "plain", "black").}

\item{font.legend}{Font settings for the legend text. Default is c(10, "plain", "black").}

\item{...}{Additional arguments passed to methods.}
}
\value{
A ggplot2 theme object customized for weighted composite endpoint survival plots.
}
\description{
Custom Theme for Weighted Composite Endpoint Survival Plots
}
\details{
This function generates a custom ggplot2 theme tailored for weighted composite endpoint survival plots, with specific font settings for various plot elements.
}
\examples{
# Example usage:
p <- ggplot(mtcars, aes(wt, mpg)) + geom_point() + theme_WCESurvZ()
print(p)
}
