% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WCE_KMSurv.R
\name{WCE_KMSurv}
\alias{WCE_KMSurv}
\title{WCE_KMSurv}
\usage{
WCE_KMSurv(
  data,
  weights = NULL,
  conf.int = 0.95,
  conf.type = "log",
  testmethod = c("logrank", "none")
)
}
\arguments{
\item{data}{A data frame containing the input data.}

\item{weights}{A named vector of weights corresponding to the events.}

\item{conf.int}{Significance level for calculating confidence intervals. Default is 0.95.}

\item{conf.type}{Type of confidence interval to calculate. Default is "log".}

\item{testmethod}{Method to use for the log-rank test. Options are "logrank" or "none". Default is "logrank".}
}
\value{
An object of class "WCEKM" containing the following elements:
\describe{
  \item{km_data}{A data frame with the combined Kaplan-Meier survival data for all groups.}
  \item{events}{A sorted vector of event names where status = 1.}
  \item{groups}{A sorted vector of group names.}
  \item{sample_size}{A named vector of the sample sizes for each group.}
  \item{followup}{The maximum follow-up time across all groups.}
  \item{pvalue}{The p-value for the log-rank test.}
  \item{testmethod}{The method used for the log-rank test.}
}
}
\description{
Generate Weighted Composite Endpoint Kaplan-Meier Survival Data
}
\details{
This function calculates weighted Kaplan-Meier survival curves for recurrent events in different groups.
}
\examples{
# Example usage:
weights <- c(CVdeath = 1, MI = 0.55, Stroke = 0.455)
WCE_obj <- WCE_KMSurv(surv_exdata, weights)

}
