% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv_exdata.R
\docType{package}
\name{WCESurvZ-package}
\alias{WCESurvZ}
\alias{WCESurvZ-package}
\title{WCESurvZ: Weighted Composite Endpoint Survival Analysis Tools}
\description{
WCESurvZ provides tools for calculating weighted survival rates and survival data based on sample events and event times. It includes functions for performing logrank tests for comparing two groups, accounting for recurrent events, and simulating data based on event distribution parameters. The package is designed to facilitate advanced survival analysis incorporating weighted composite endpoints.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/WCESurvZ}
}

}
\keyword{internal}
