% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logrankTest.R
\name{logrankTest}
\alias{logrankTest}
\title{logrankTest}
\usage{
logrankTest(data)
}
\arguments{
\item{data}{}
}
\value{
NA
}
\description{
Perform Log-Rank Test for Survival Data
}
\details{
This function performs a log-rank test to compare survival curves between different groups.
}
\examples{
NA
}
