% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.WCEKM.R
\name{plot.WCEKM}
\alias{plot.WCEKM}
\title{Plot Weighted Composite Endpoint Kaplan-Meier Survival Curves}
\usage{
\method{plot}{WCEKM}(
  WCEKM_obj,
  x_limits = NULL,
  y_limits = c(0, 1),
  title = NULL,
  base_size = 12,
  x_label = NULL,
  y_label = NULL,
  x_ticks = NULL,
  y_ticks = NULL,
  colors = NULL,
  legend_position = c(0.9, 0.8),
  legend_title = NULL,
  legend_labels = NULL,
  theme = theme_WCESurvZ(base_size = base_size)
)
}
\arguments{
\item{WCEKM_obj}{An object of class "WCEKM" containing Kaplan-Meier survival data.}

\item{x_limits}{A vector specifying the limits for the x-axis. Default is NULL.}

\item{y_limits}{A vector specifying the limits for the y-axis. Default is c(0, 1).}

\item{title}{The title of the plot. Default is "非加权Kaplan-Meier生存曲线".}

\item{x_label}{The label for the x-axis. Default is NULL.}

\item{y_label}{The label for the y-axis. Default is NULL.}

\item{x_ticks}{A vector specifying the ticks for the x-axis. Default is NULL.}

\item{y_ticks}{A vector specifying the ticks for the y-axis. Default is NULL.}

\item{colors}{A vector of colors for the different groups. Default is NULL.}

\item{legend_position}{A vector specifying the position of the legend. Default is c(0.9, 0.8).}

\item{legend_title}{The title for the legend. Default is NULL.}

\item{legend_labels}{A vector of labels for the legend. Default is NULL.}

\item{font_size}{The base font size for the plot. Default is 12.}
}
\value{
A ggplot2 object showing the Kaplan-Meier survival curves.
}
\description{
This function plots Kaplan-Meier survival curves for weighted composite endpoints.
}
\examples{
# Example usage:
weights <- c(CVdeath = 1, MI = 0.55, Stroke = 0.455)
WCE_obj <- WCE_KMSurv(surv_exdata, weights)
plot(WCE_obj)

}
