% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_WCESurvZ.R
\name{theme_WCESurvZ}
\alias{theme_WCESurvZ}
\title{theme_WCESurvZ}
\usage{
theme_WCESurvZ(
  base_size = 12,
  base_family = "",
  font.main = c(16, "plain", "black"),
  font.submain = c(15, "plain", "black"),
  font.x = c(14, "plain", "black"),
  font.y = c(14, "plain", "black"),
  font.caption = c(15, "plain", "black"),
  font.tickslab = c(12, "plain", "black"),
  font.legend = c(10, "plain", "black"),
  ...
)
}
\arguments{
\item{base_size}{NA}

\item{base_family}{NA}

\item{font.main}{NA}

\item{font.submain}{NA}

\item{font.x}{NA}

\item{font.y}{NA}

\item{font.caption}{NA}

\item{font.tickslab}{NA}

\item{font.legend}{NA}

\item{...}{NA}

\item{legend}{NA}
}
\value{
NA
}
\description{
theme_WCESurvZ
}
